added flavor uncertainty to tau lepton SF

+++ include/identifier.h
@@ -43,9 +43,9 @@ public:
     }
         
 private:
-
-    int32_t id_;
-
+   
+   int64_t id_;
+   
 };
